phidata/vediosummarizer/requirements.txt
phidata
google-generativeai
duckduckgo-search
streamlit==0.62.0
